Fix #44 Abstract class removal, metricsView directly extending ViewPart
